Step1: DateFormatter 싱글톤으로 분리
