+++ WeatherForecast/WeatherForecast/Resource/SceneDelegate.swift
@@ -16,8 +16,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window = UIWindow(windowScene: scene)
         
         let viewController = WeatherViewController()
-        let nvc = UINavigationController(rootViewController: viewController)
-        window?.rootViewController = nvc
+        let navigationController = UINavigationController(rootViewController: viewController)
+        window?.rootViewController = navigationController
         window?.makeKeyAndVisible()
     }
 }
